+++ src/IRremoteInt.h
@@ -163,7 +163,7 @@ struct IRData {
     decode_type_t protocol;     ///< UNKNOWN, NEC, SONY, RC5, ...
     uint16_t address;           ///< Decoded address
     uint16_t command;           ///< Decoded command
-    uint16_t extra;             ///< Used by MagiQuest and for Kaseikyo unknown vendor ID
+    uint16_t extra;             ///< Used by MagiQuest and for Kaseikyo unknown vendor ID.  Ticks used for decoding Distance protocol.
     uint8_t numberOfBits; ///< Number of bits received for data (address + command + parity) - to determine protocol length if different length are possible.
     uint8_t flags;              ///< See IRDATA_FLAGS_* definitions above
     uint32_t decodedRawData;    ///< Up to 32 bit decoded raw data, used for sendRaw functions.
@@ -258,6 +258,8 @@ public:
     bool decodeSamsung();
     bool decodeSharp(); // redirected to decodeDenon()
     bool decodeSony();
+
+    bool decodeDistance();
 
     bool decodeHash();