+++ changelog.md
@@ -3,6 +3,7 @@ The latest version may not be released!
 ## 3.2.1
 - Fix errors if LED_BUILTIN is not defined.
 - Fixed error for AVR timer1. Thanks to alexbarcelo.
+- New example IRremoteExtensionTest.
 
 ## 3.2.0
 - Fix for ESP32 send Error, removed `USE_SOFT_SEND_PWM` macro.