+++ changelog.md
@@ -1,10 +1,11 @@
 # Changelog
 The latest version may not be released!
-## 3.2.1
+## 3.3.0
 - Fix errors if LED_BUILTIN is not defined.
 - Fixed error for AVR timer1. Thanks to alexbarcelo.
 - New example IRremoteExtensionTest.
 - Enabled megaAVR 0-series devices.
+- Added universal decoder for pulse width or pulse distance protocols.
 
 ## 3.2.0
 - Fix for ESP32 send Error, removed `USE_SOFT_SEND_PWM` macro.